% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_haplotypes.R
\name{haplotype_counts}
\alias{haplotype_counts}
\title{Count the number of times each haplotype occurs}
\usage{
haplotype_counts(x)
}
\arguments{
\item{x}{DataFrame (by locus) of character vectors containing haplotypes (rows) where alleles are separated by comma's, e.g. "13,14.2" is a haplotype}
}
\value{
Integer vector
}
\description{
Count the number of times each haplotype occurs
}
