% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_haplotypes.R
\name{haplotype_table}
\alias{haplotype_table}
\title{List unique haplotypes with their counts}
\usage{
haplotype_table(x)
}
\arguments{
\item{x}{DataFrame (by locus) of character vectors containing haplotypes (rows) where alleles are separated by comma's, e.g. "13,14.2" is a haplotype}
}
\value{
DataFrame
}
\description{
List unique haplotypes with their counts
}
